Move remaining pipeline inputs to fit new structure  (#1182)
