+++ compositor_pipeline/src/pipeline/channel/mod.rs
@@ -1,0 +1,7 @@
+mod encoded_data_output;
+mod raw_data_input;
+mod raw_data_output;
+
+pub use encoded_data_output::EncodedDataOutput;
+pub use raw_data_input::RawDataInput;
+pub use raw_data_output::RawDataOutput;
