+++ compositor_pipeline/src/pipeline/rtmp/mod.rs
@@ -1,0 +1,3 @@
+mod rtmp_output;
+
+pub use rtmp_output::RtmpClientOutput;
