+++ compositor_pipeline/src/pipeline/rtmp/rtmp_output.rs
@@ -4,18 +4,20 @@ use crossbeam_channel::{bounded, Receiver, Sender};
 use ffmpeg_next::{self as ffmpeg, Rational, Rescale};
 use tracing::{debug, error};
 
-use crate::prelude::*;
 use crate::{
     event::Event,
-    pipeline::encoder::{
-        encoder_thread_audio::{spawn_audio_encoder_thread, AudioEncoderThreadHandle},
-        encoder_thread_video::{spawn_video_encoder_thread, VideoEncoderThreadHandle},
-        fdk_aac::FdkAacEncoder,
-        ffmpeg_h264::FfmpegH264Encoder,
+    pipeline::{
+        encoder::{
+            encoder_thread_audio::{spawn_audio_encoder_thread, AudioEncoderThreadHandle},
+            encoder_thread_video::{spawn_video_encoder_thread, VideoEncoderThreadHandle},
+            fdk_aac::FdkAacEncoder,
+            ffmpeg_h264::FfmpegH264Encoder,
+        },
+        output::{Output, OutputAudio, OutputVideo},
     },
 };
 
-use super::{Output, OutputAudio, OutputVideo};
+use crate::prelude::*;
 
 #[derive(Debug, Clone)]
 struct Stream {
